fixed bug on block cell and compute max label on whole data

+++ src/embdevtools/celltrack/core/tools/tools.py
@@ -95,7 +95,7 @@ def check_and_fill_error_correction_args(error_correction_args):
 
 # @njit()
 def increase_point_resolution(outline, min_outline_length):
-    rounds = np.ceil(np.log2(min_outline_length / len(outline)))
+    rounds = int(np.ceil(np.log2(min_outline_length / len(outline))))
     if rounds <= 0:
         newoutline_new = np.copy(outline)
     for r in range(rounds):
